# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API MedicAid",
        "description": "Documentación de la API de MedicAid (Usuarios, Medicamentos, Tratamientos, Historiales y Auditoría)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/audit-logs": {
            "get": {
                "tags": [
                    "Audit Logs"
                ],
                "summary": "Listar registros de auditoría",
                "operationId": "8a24c03b06fe3071c4035f4176892716",
                "responses": {
                    "200": {
                        "description": "Listado paginado de registros de auditoría"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Audit Logs"
                ],
                "summary": "Crear registro de auditoría",
                "operationId": "ea7ff7fb9a702bd6cf96e05245992a75",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "usuario_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "accion": {
                                        "type": "string",
                                        "example": "CREAR_TRATAMIENTO"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "El usuario creó un nuevo tratamiento"
                                    },
                                    "ip_address": {
                                        "type": "string",
                                        "example": "127.0.0.1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registro de auditoría creado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/audit-logs/{id}": {
            "get": {
                "tags": [
                    "Audit Logs"
                ],
                "summary": "Obtener un registro de auditoría",
                "operationId": "b66779a33559eaaaad6e147d5e916f1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registro de auditoría encontrado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Audit Logs"
                ],
                "summary": "Eliminar registro de auditoría",
                "operationId": "9e05da700d0bf8d60c2f355e6e438771",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/historial": {
            "get": {
                "tags": [
                    "Historial"
                ],
                "summary": "Listar historiales",
                "operationId": "1ba06fa09b858db7daee2bbd3ac32fa6",
                "responses": {
                    "200": {
                        "description": "Listado de historiales con sus tratamientos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Historial"
                ],
                "summary": "Crear historial",
                "operationId": "3c7fb2088a76a1fa0c1cb8a5571ed40f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tratamiento_id",
                                    "fecha_hora"
                                ],
                                "properties": {
                                    "tratamiento_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha_hora": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-11-24 10:30:00"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "pendiente"
                                    },
                                    "notas": {
                                        "type": "string",
                                        "example": "Tomar en ayunas"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Historial creado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/historial/{historial}": {
            "get": {
                "tags": [
                    "Historial"
                ],
                "summary": "Obtener un historial",
                "operationId": "c82e05d1e1ad3c96f40ec003d7a900e6",
                "parameters": [
                    {
                        "name": "historial",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historial encontrado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Historial"
                ],
                "summary": "Actualizar historial",
                "operationId": "fe8169d7f6111918b12cb5d32afe3fd8",
                "parameters": [
                    {
                        "name": "historial",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tratamiento_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha_hora": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "tomada"
                                    },
                                    "notas": {
                                        "type": "string",
                                        "example": "Sin efectos secundarios"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Historial actualizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Historial"
                ],
                "summary": "Eliminar historial",
                "operationId": "42d6c66789e5095239097c15873b304b",
                "parameters": [
                    {
                        "name": "historial",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}